refactor : 코드 정리, 폴더 변경, 삭제
