+++ src/test/java/com/ezfarm/ezfarmback/auth/TokenTest.java
@@ -1,5 +1,6 @@
-package com.ezfarm.ezfarmback.security;
+package com.ezfarm.ezfarmback.auth;
 
+import com.ezfarm.ezfarmback.common.CommonSecurityTest;
 import com.ezfarm.ezfarmback.config.security.AppProperties;
 import com.ezfarm.ezfarmback.security.local.TokenProvider;
 import com.ezfarm.ezfarmback.user.dto.AuthResponse;
